pyproject.toml
[tool.poetry]
name = "project-online-electronics-trading-network-platform"
version = "0.1.0"
description = ""
authors = ["AlexandraAvtaykina <justalexandra@mail.ru>"]
readme = "README.md"
packages = [{include = "project_online_electronics_trading_network_platform"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
ipython = "^8.22.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
django-filter = "^24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
